requirements.txt
# Python dependencies for LeetCode Daily Bot
playwright>=1.40.0
playwright-stealth>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot==13.15
